move Location to its own file
