+++ src/common/location.ts
@@ -1,0 +1,6 @@
+
+export interface Location {
+    x: number;
+    y: number;
+    z: number;
+}
