+++ src/server/connection-server.ts
@@ -1,6 +1,7 @@
 import http from 'http';
 import { Socket } from 'socket.io';
-import { ItemState, Location } from 'src/common/item';
+import { ItemState } from 'src/common/item';
+import { Location } from "src/common/location";
 import { EVENTS } from '../common/events';
 import { ServerEntity, ServerEntityProperties } from './entities/server-entity';
 import { EntityServer, EntityServerTemplate, serializeCaveItems } from './entity-server';
