+++ src/server/entity-server.ts
@@ -2,7 +2,8 @@ import axios from 'axios';
 import { MapState } from '../common/map';
 import { Tile } from '../common/tile';
 import { EVENTS } from '../common/events';
-import { Item, ItemState, Location } from '../common/item';
+import { Item, ItemState } from '../common/item';
+import { Location } from "../common/location";
 import { DIRS, getMovement, Movement } from '../common/movement';
 import { Cave, CaveItems, CaveTemplate } from './cave';
 import { ServerEntity, ServerEntityProperties } from './entities/server-entity';
